+++ CHANGES.md
@@ -2,12 +2,14 @@
 
 3.2.0 (pre-release)
 
+* #162 - LoggingFilterSwitch support
 * #202 - added support to AuditTo.Logger
 * #203 - added support for custom types in arrays and custom collections
 * #218 - fixed an issue with `dotnet restore` with `rid` specified if referenced from `netstandard` project
 * #219 - reduced search graph for configuration dlls to avoid native assets
 * #221 - added support for conditional/leveled enrichers from Serilog 2.9+
 * #222 - updated Microsoft.Extensions.DependencyModel
+* #231 - make '$' sign optional for minimum level / filter switch declarations
 * #237 - DependencyContextAssemblyFinder fix: check `serilog` at the start of the name for any dependent package
 * #239 - handle NotSupportedException for .net 5.0 single file applications